Update `If` regressions
These test cases show another interesting issue we are currently facing:
When using an `if`, the boolean condition variable that decides on the
branch needs to be input into two separate operators, one for sending
the data into the proper branch and one for receiving the data from the
correct branch. As of now, this is not created correctly. We need a
dispatch for the bool variable (it doesn't even need cloning due to
`bool` implementing `Copy`)