filtered input code, return result, fix import
First:
When building modules for each parallelized function
we used to transfer the original code to the new modules
to preserve imports and definitions of variables and classes.
However function definitions need not, and 'if '__name__'== "__main__"'
must not be transfered. SO the code is filtered for those statements now.

Second:
The code of compiled functions was replaced by a call to the according
parallel module. However I forgot to return the result of this call.
This is fixed now

Third:
I used 'from . import algo as algo_parallel' before, this only works if
I put all produced modules in a package. I have to evaluate this before, so
now I use an ordinary  'import algo as algo_parallel'