makeshift solution for M3 Channels
Some patterns for channel, receive and send
conversion where not supported but resulted from
compiling test code. They are, for now supported the
same way as in STM, or by placeholder type (!) respectively.
This is no working permanent solution, but a makeshift to proceed
with fixing logic in the front and middle end.