infer from local vars and literals
To replace, or at least augment type extraction from
files in scope I start inferring function type from
local context.
With this commit, this approach works for literals
and locally defined variables. It requires the results of
each local binding statement to be type annotated i.e.

``let x:sometype = somefunction();``

The loadTypes step is replaced returning the namespace as is.
Next step is to bring the typed parameters of the parsed algorithms
into the conversion context.