thread context through conversion
To also get the algo parameters into the
typing context this commit threads the context monad
through the whole conversion (including the Rust->Subset
conversion).

An alternative would be to have the argument conversion
return a HM instead of just the converted args. This
HM could be fed into later steps as initial context,
instead of the previous HM.empty