test enforcement of progr. model
So far we basically had 'positive' tests,
i.e. do things we want to work actually work.
However our programming model has assumptions on
things that should not work, while they might be
valid in the input language -> We need tests that
expect compiler errors.

Rust is not a perfect example, because the borrow
checker plays the bouncer for us. However a) we also
(want to) support other languages as well and b) I might
just not be aware of how our progr. model assumptions
can be cleverly violated in Rust.

So -> We need test and here we go.