add returnType field
As long as we do not type all expressions,
we need at least the return type of each algorithm
for type propagation.
So this commit threads this new field through the backend.

Also I renamed some fields in the core to better distiguish
e.g. NS (now HostModule) from Namespace (the 'working representation'
of the namespace)