recurFun and return typing
The return value i.e. the Var expression
terminating each algo had no type parameter so we couldn't
propagate it's type to the backend although we knew it.
Hence this expression is now ´Var bnd ty´ to transfer the type
properly.

Also this might fix type propagation for recursive functions.