adding type to internal var representation
To type the added nodes correctly we currently lack some information at generation time and need to add
typing to those nodes later on in the type propagation. This actually means we use information about variable types to derive and annotate function types, then we throw away the varaible types, then we cannot type the added nodes, because we don't have types on the variables and then we extract type info from the annotated functions and try to propagate them to untyped variables.

Therefor I'd like to add type annotation to the variables directly.