typing vars in python
As we do not need typing for the Python backend, we do not need to extract types from the code, and hence do
not need to change the subset language to store extracted variable types. Instead we just lower to the
Frontend Language (now having VarE/VarP typed) by adding the default type PythonObject.