wip Rust Subset remains Maybe typed
WIP -> Does not compile

We might want to change the source of type annotations for
variables in the Frontend Language. Also, even if we take types directly
from variable annotation, we only know them from the context.
Hence the conversion from Rust to the supported subset has still
(Maybe Type)ed Vars and Pats, and we need to introduce them in the
conversion from Subset to Frontend Language.