wip typed vars in Alang + DFLang
Now typed vars are part of Alang i.e. Variables, Lambda abstractions and Let-bound patterns take a `TypedBinding = TBind Binding ArgType` instead of just a binding.
Next step is to integrate that information in the lowering from ALang to DFLang and finally we will hopefully be able to exploit the types to correctly
assign types during transformations in the compiler and eliminate the need for TypePropagation.