typed Bindings in DFLang
For now this just threads typed bindings (TBind Binding Argtype) through where we used to have only bindings.
We should be able to later differentiate TBinds further, so that we can actually index the function types directly with the argument types.
i.e. we can have typed bindings directly representing Objects and type functions as acting on objects instead of having an additional marker type for the
function.