Fixed equality
As also noted in the code, we currently do not make sure that everything is correctly typed all the time. This means a typed binding can have different annotations at binding and usage side (namely the correct type and TypeVar in cases where we do not infer properly right now). Thats why we need to compare only the bindings for now, because if we compared binding+type the transformation fitlering dead code finds no usage fr currently misstyped usage sides and removes code that should not be removed.