reactivate Rust Typeextraction
We want to add type extraction again a source of type information. However, as it still has some (implementation or general) shortcommings, we now want it to be the
an optional source of type information and require annotation if it fails.

So the first steps here are to a) reactivate the code and b) make finding the imported libraries optional. For now there's no particular failure mechanism. If an imported file, or in fact the compiled moudle is not found, we just return an empty Rust module.