issue resolving 'crate::' imports
To resolve imports we can not just translate the path to files as the example of "crate::" imports shows. There is no directory 'crate/' to search files in.
However as this is the most common case we'll need for testing, I'll fix this concrete problem temporarily by replacing 'crate' with '.' assuming trivial crate imports from the same directory.