handle global definitions
Global definitions are acceptable as constant. To include them, we need to tell the compile that they are (implicit) arguments to the algorithms.
To teach this to the compiler we need to make them actual arguments, just like the original argumetns of the algorithms.
We will however, need to treat those kinds of arguments differently in the backend. Hence the shortcut for now, is adding names of global vars to namespace,so that we can
(except for cases of name shaddowing) distinguish explicit from implicit args in the backens again.