rename and type UnitP to WildP
Actually patterns like in 'let p = ' or 'for p in ...' are never actually unit (e.g. ()),
if they are not used they are '_' mostly. Hence they need another name and they need to have a type just for the
purpose of keeping types correct i.e. let _::pty = term::pty