wip fix Python errors after typing
This fixed the problem of compilation getting stuck n ALang passes. Problem was, that the pass assumed all vars to be bound in scope so function names of the
host language must be represented as LitE not as VarE. Also the added unit literal in fucntion that do not take parametersis only filtered
out in the backend integration. This was somehow introduced by typing everything (!?) and should be moved into the comomn backend as soon as possible.