fix endless loop problem
The problem was, that a recursive rewrite checks if its done, by comparing its input to its output.
If they are equal, the rewrite returns. This failed (i.e. looped forever) if something had the
VarType: TypeFunction, because Eq in this case fell into the default pattern, which returned False.