hack-fix for maxType
Deciding what the more specific type of two given candidates is or if there should be a type error
requires some understanding of the host types we currently haven'tbuild in. In particular if we assign literals
internal types as TypeNat and TypeBool, we currently have no way to tell if a given host type annotation
is coercible to those types or not. Likewise we don't handle generics and we need a rule for
"concrete struct type" > self.