wip fix return type propagation
This commit removes the previouse version of type propagation. Also the funcions returning the types of expression and literals now
return the actual types and not, as before only the return types in in case of abstraction terms.

This still causes problems though and not all basic tests for Shared Memory pass