wip: add return type again
I think the typeSystem should check if every expression is well formaed and if it's type matches the expected return type. We know that type from the algo extration.
However I'm currently not sure if it's better to have that type separately, or to actually translate every algorithm to be let abig let expression with the outermost
bound variable being the algo itself. This seems to happen for algos, that are called by other algos, but not for the ones that aren't.