temporary fix for dependency issue
There was a parse error whith the current resolver+language-rust in the newer version so I switched back for now.
In the long run we need to check whether to fork and maintain our own version.

Also tomland requried a lower version of base than the current compiler -> complying to that requirement causes an avalannge
of other dependency issuues so I commented out the rust-macro-support info for now.