adding unresolved LamE and AppE
We have the following problem:
Inside Ohua, patterns and arugments must not be empty, while in the input code they might be.
If the parameters of a lambda or the arguments of an application are empty, a unit argument has to be added.
This shouldn't happen in the language integrations, as it is a "compiler thing". Therefor the integrations
produce "unresolved" versions of lambda and app now, that allow empty patterns and are later converted augmented with unit args/patterns before type check.