wip typing, stuck at stateful
The current problem is, that we need the fully qualified type of an object,
to correctly type the methods called on it, because the type of the object proved the namespace
of the method.

This commit adds a "manipulation" of the namespace of the method, adding the objects type to it.
It only works for methods that are function literals currently.
Still the problem is, that e.g. for an object created via std::sync::Arc::new(), we only get
Arc as the type and cannot type Arc::clone without the remaining namespace refs.