add lambda to let transf again
we used to wrap algo bodies in a lambda term to capture the arguments, but transform this to
let biding expressions for every original argument (EnvLits from compiler perspective) to local variables.

While the plan is to remove the need fro this, by kepping the env vars in a context and thread them to the compiler,
we currently need to kepp this informatin and hence the transformation in the algo bodies as we process them.