temp fix algo types
We need the types of the algos, i.e. their decalred argument and return types for the type check obviously.
Currently we pass algos as lambda abstractions to the type system. So while we usually don't use patterns for typing, in case of the
outermost expression i.e. the algo itsef, we need to take the pattern types and names to add
them to the context gamma to type the body of the algos.