fix algo inlining
Algo inlining used to search only for algos represented as function literals with function types.
We could expect integrations to ensure that, but as we know the names of algorithms, we should
(in the interest of robustness) also accept algorithms represented as variables and having an unknown
(mostly TStar) type. This is hereby done (and fixes some test cases at least in this regard)