fix namespace local
When we extract funtion types in Rust, the namespaces are taken from file path's.
This works for imported files as they are taken relative to the compiled file,
but not for the compiled file itself because it introduces an actual (temp)
system filepath in the anemspace, which we cannot identify/map during type checking.