merge gammas, "interim" arg type equality
We typecheck bottom up, this means, variables are typed where they are used and in case of
let bindings where they are produced. This means, we need to merge the type information
in let expression Let e = e1 in e2 from e1 and e2, and in case of Applications we need to merge the arg
type information to a common context.

Also we actually need to compare the argument types comming from function declarations to
the types infered for the arguments. We current cannot do this correctly, because we
a) dont enforce a subtyping and generics comparison operator of HostType (which means they'd
need to be stringly equal) and b) we still have a problem with literal types.
At least the present "interim" comparison allows to distinguish a function that should have no arguments
i.e. it's argtype is Unit, from a function that actually has one argument.