replace Expr Type, unwrap Algos
This replaces the Frontend Expression type ba the new parameterized version. It also adds a function to "unwrap" algorithms before
lowering them to ALang. The reason is, that ALang expects Algos to be Let expressions, while in the frontend they have an
"outermost layer" that is a lamda, because this allows us to carry the type information of the input parameters to the
typecheck.