fix typing with empty args lists
Now we do not assume that functions have a unit arg upon typing or fix it during typing.
So we need to handle empty argument lists and empty pattern lists special and we need
to check for one element lists in the input types of functions if that one element
is in fact the internal unit type, which means that type checking shoud expect no arguments when
respective functions are called.

The later is obviously error prone. We need to have a clear separaration between unit functions and functions with
one (actual) argument.