don't accept partial application
When type cecking function applications, we used to check only, if there are no mroe arguments than decalred parameter types.
If there where less arguments than declared parameters, we handled this as partial application and the return type
of the application would be a function type. However, our host languages (Rust and Python) do not support partical application,
and supporting it let to type errors downstream, that where hard to trace back to the original point of wrong function calls.