type pattern by annotation
When patterns are assigned but not used, they wont appear in the resolved context and
typing used to fail at this point (saying the pattern was deleted from the context, while it just wasn't added
upon typing it's usage). As this might happen e.g. in loop patterns, we'll take the type annotation e.g. gathered from the return
type of the generator, also as valid type source for patterns now.