interim fix seq looping
receiving tasks for seq'ed operations in oops need to be loopeed as well, so for now we'll do this.
still it's just a hotfix. in the long term, what is "seq'ed" in the frontend should be "unseq'ed" in the backend,
meaning Seq should become an expression in TCLang such that seq nodes can be fused with the tasks that produce the "ignored" values.
see issue https://github.com/ohua-lang/ohuac/issues/42