new FunType
Function types use to have non-empty argument lists, which we enforced by typing f() as F [Unit]. However Unit was an internal, resoved type so we
a) couldn't allready type those function in unsresoved expresssions (wich we need to do for Python) and
b) needed to make a pattern matching on the first argument type to discriminate functions with one (actual) argument from functions
we internally augmented with a unit.

This adds some helper functions to retrieve expected argument types from function types.