parameterize FRLang by EmbExpr
Make the type of embedded expressions a parameter of the frontend language and of
frontend integrations. For instance in Python we (may) want embedded expressions to be
Python expressions, or 'suits' so we do not want to tie the type of embedded expressions to the type
of language expressions (Expr lang). Also we want to make the type f embedded expressions consistent for frontend
and backend integrations I assume.