wip call annotations
currently in progress is support in the intergations and frontend language. missing piece
(and general pain point to be removed by rewrite) is the show constraint on the annotations.

After that, annotations need to be
a) added to StateFunE's in the frontend because method calls are not directly
   translated to apply expressions (which are annotated in FrLang now)
b) added to ALang
c) added to DFLang
d) added to TCLang
to conclude the functional part

Afterwards we should consider summarizing the expression type, type of embedded expressions and type of annotations
because type declarations become unreadable currently